refactor(suite): extract AddAccountBannerAboutNetworks
